+++ application/Bootstrap.php
@@ -23,29 +23,7 @@ class Bootstrap extends Zend_Application_Bootstrap_Bootstrap
 
         return parent::run();
     }
-
-    /**
-     * Initialisation du cache APC
-     */
-    protected function _initCache()
-    {
-        return Zend_Cache::factory('Core', 'APC', array(
-            'lifetime' => getenv('PREVARISC_CACHE_LIFETIME'),
-            'cache_id_prefix' => 'prevarisc'
-        ));
-    }
-
-    /**
-     * Initialisation du cache APC spécial recherches
-     */
-    protected function _initCacheSearch()
-    {
-        return Zend_Cache::factory('Core', 'APC', array(
-            'lifetime' => getenv('PREVARISC_CACHE_LIFETIME'),
-            'cache_id_prefix' => 'prevarisc_search'
-        ));
-    }
-
+    
     /**
      * Initialisation de l'auto-loader
      */
@@ -54,11 +32,71 @@ class Bootstrap extends Zend_Application_Bootstrap_Bootstrap
         $autoloader = Zend_Loader_Autoloader::getInstance();
 
         $autoloader_application = new Zend_Application_Module_Autoloader(array('basePath' => APPLICATION_PATH, 'namespace'  => null));
-
+        
+        $autoloader_application->addResourceType('cache', 'cache/', 'Cache');
+        
         $autoloader->pushAutoloader($autoloader_application);
 
         return $autoloader;
     }
+    
+    /**
+     * Initialisation d'un cache standard
+     * @param array $frontendOptions surcharge des options de configuration du front
+     * @param array $backendOptions surcharge des options de configuration du back
+     * @return Cache une instance de cache
+     */
+    protected function getCache(array $frontendOptions = array(), array $backendOptions = array()) {
+        
+        $options = $this->getOption('cache');
+        
+        return Zend_Cache::factory(
+                // front adapter
+                'Core',
+                // back adapter
+                $options['adapter'], 
+                // frontend options
+                array_merge(array(
+                    'caching'  => $options['enabled'],
+                    'lifetime' => $options['lifetime'],
+                    'cache_id_prefix' => 'prevarisc_',
+                    'write_control' => $options['write_control'],
+                ), $frontendOptions),
+                // backend options
+                array_merge(array(
+                    'servers' => array(
+                        array(
+                            'host' => $options['host'], 
+                            'port' => $options['port'],
+                        ),
+                    ),
+                    'compression' => $options['compression'],
+                ), $backendOptions),
+                // use a custom name for front
+                false,
+                // use a custom name for back
+                $options['customAdapter'],
+                // use application's autoload if an adapter is not loaded
+                true);
+    }
+
+    /**
+     * Initialisation du cache objet de l'application
+     */
+    protected function _initCache()
+    {
+        return $this->getCache();
+    }
+
+    /**
+     * Initialisation du cache spécial recherches
+     */
+    protected function _initCacheSearch()
+    {
+        return $this->getCache();
+    }
+
+    
 
     /**
      * Initialisation de la vue